Renamed arguments.py to scons_args.py.
