+++ apio/managers/scons.py
@@ -22,8 +22,8 @@ import semantic_version
 
 from apio import util
 from apio import pkg_util
-from apio.managers.arguments import process_arguments
-from apio.managers.arguments import serialize_scons_flags
+from apio.managers.scons_args import process_arguments
+from apio.managers.scons_args import serialize_scons_flags
 from apio.managers.system import System
 from apio.apio_context import ApioContext
 from apio.managers.project import Project
